Support arm7a targets and enforce ILP32 safety
Refactor hardware-dependent routines for environments lacking
native 128-bit integers by providing a 64-bit signed multiplication
fallback and a lightweight Xorshift64 PRNG. Syscall definitions are
also expanded to cover standard ARM EABI targets.

To guarantee deterministic behavior across ILP32 and LP64 data
models, ambiguous 'long' types in sysfs parsing are strictly bounded
to int32_t. Logging formats are fortified with explicit casts, ensuring
cross-architecture stability without introducing runtime overhead on AArch64.

+++ src/sys/rlimit.c
@@ -19,8 +19,8 @@ int pg_rlimit_set_max_fd(void)
 		return -err;
 	}
 
-	LOGD("rlimit: fd limit maximized to %lu",
-	     (unsigned long)rl_fd.rlim_cur);
+	LOGD("rlimit: fd limit maximized to %llu",
+	     (unsigned long long)rl_fd.rlim_cur);
 	return 0;
 }
 
@@ -41,7 +41,7 @@ int pg_rlimit_set_stack(size_t bytes)
 		return -err;
 	}
 
-	LOGD("rlimit: stack bounded to %lu bytes",
-	     (unsigned long)rl_stack.rlim_cur);
+	LOGD("rlimit: stack bounded to %llu bytes",
+	     (unsigned long long)rl_stack.rlim_cur);
 	return 0;
 }
